docker/api-server/requirements.txt
fastapi
uvicorn[standard]
pyjwt
pyjwt[crypto]
pydantic[email]
pydantic-settings
requests